[rename] Portal -> portals
